refactor: Article API 도메인으로 이동
