+++ backend/src/main/java/com/icehufs/icebreaker/service/implement/AuthServiceImplement.java
@@ -13,7 +13,7 @@ import com.icehufs.icebreaker.common.CertificationNumber;
 import com.icehufs.icebreaker.dto.request.auth.*;
 import com.icehufs.icebreaker.dto.response.ResponseDto;
 import com.icehufs.icebreaker.dto.response.auth.*;
-import com.icehufs.icebreaker.entity.Article;
+import com.icehufs.icebreaker.domain.article.domain.entity.Article;
 import com.icehufs.icebreaker.entity.AuthorityEntity;
 import com.icehufs.icebreaker.entity.BanDurationEnum;
 import com.icehufs.icebreaker.entity.BanReasonEnum;
@@ -22,7 +22,7 @@ import com.icehufs.icebreaker.entity.User;
 import com.icehufs.icebreaker.entity.UserBan;
 import com.icehufs.icebreaker.provider.EmailProvider;
 import com.icehufs.icebreaker.provider.JwtProvider;
-import com.icehufs.icebreaker.repository.ArticleRepository;
+import com.icehufs.icebreaker.domain.article.repository.ArticleRepository;
 import com.icehufs.icebreaker.repository.AuthorityRepository;
 import com.icehufs.icebreaker.repository.CertificationRepository;
 import com.icehufs.icebreaker.repository.UserBanRepository;
